Extract SiteTitle component
And allow it to center/truncate correctly
